+++ src/lib/forms-schema/elements/LocationElement.ts
@@ -31,9 +31,4 @@ export default Joi.object({
     then: Joi.string().uuid().required(),
     otherwise: Joi.any().strip(),
   }),
-  integrationEnvironmentId: Joi.when('showStreetAddress', {
-    is: true,
-    then: Joi.string().uuid().required(),
-    otherwise: Joi.any().strip(),
-  }),
 })